trusted-types/block-string-assignment-to-Element-setAttributeNode.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<meta http-equiv="Content-Security-Policy" content="require-trusted-types-for 'script';">
<title>trusted-types (TT): setAttributeNode with node from non-TT realm</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="nonSVGTestElements">
    <iframe srcdoc="v"></iframe>
    <embed src="v">
    <script src="v"></script>
    <object data="v"></object>
    <object codebase="v"></object>
</div>
<svg id="svgTestElements">
    <script href="v"></script>
    <script xlink:href="v"></script>
</svg>
<script>
    const passThroughPolicy = trustedTypes.createPolicy("passThrough", { createHTML: s => s });

    function runTest(aTestElement) {
        const testAttr = aTestElement.attributes[0];

        async_test(t => {
            const sourceFrame = document.createElement("iframe");

            // The markup requires the parent element to ensure the attribute is associated with the
            // correct namespace.
            sourceFrame.srcdoc = passThroughPolicy.createHTML(
            `<!DOCTYPE html>
            <head>
            <meta charset="utf-8">
            </head>
            <body>
                <` + aTestElement.parentElement.localName + `>
                    <` + aTestElement.localName + ` ` + testAttr.name + `="` + testAttr.value + `">
                    </` + aTestElement.localName + `>
                </` + aTestElement.parentElement.localName + `>
            doc without TT CSP.
            </body>`);

            t.add_cleanup(() => {
                sourceFrame.remove();
            });

            sourceFrame.addEventListener("load", t.step_func_done(() => {
                // A window is a global object which has 1-to-1 mapping to a realm, see the first
                // note of <https://html.spec.whatwg.org/#realms-settings-objects-global-objects>
                // and its following paragraph.
                assert_equals(aTestElement.ownerDocument, document);
                assert_not_equals(aTestElement.ownerDocument.defaultView, sourceFrame.contentWindow,
                "The source frame's realm differs from the test element's realm.");

                const sourceElement = sourceFrame.contentDocument.body.querySelector(aTestElement.localName);
                const sourceAttr = sourceElement.getAttributeNode(testAttr.name);
                sourceElement.removeAttributeNode(sourceAttr);

                assert_throws_js(TypeError, () => { aTestElement.setAttributeNode(sourceAttr); });
            }));

            document.body.append(sourceFrame);

        }, `Importing a "` + testAttr.name + `" attribute node created in a non-TT enforcing ` +
        `realm to a TT enforcing realm to a "`+ aTestElement.localName + `" node with parent="` +
        aTestElement.parentElement.localName + `" throws.`);
    }

    for (const testElement of document.querySelectorAll("#nonSVGTestElements *")) {
        runTest(testElement);
    }

    for (const testElement of document.querySelectorAll("#svgTestElements *")) {
        runTest(testElement);
    }
</script>
</body>
</html>
